+++ packages/undercut/pull.js
@@ -6,5 +6,6 @@ export * from "./src/pull/pull_sources.js";
 export * from "./src/pull/pull_targets.js";
 
 export { compare };
+export * from "./src/utils/coroutine.js";
 export * from "./src/utils/iterable.js";
 export { asc, desc } from "./src/utils/ordering.js";