+++ packages/undercut/src/push/operations/sort.js
@@ -1,42 +1,40 @@
 import { assert, assertFunctor } from "../../utils/assert.js";
 import { numbers, strings } from "../../utils/compare.js";
-import { closeObserver } from "../../utils/observer.js";
+import { abort, asObserver, close, Cohort } from "../../utils/coroutine.js";
 import { asc, desc } from "../../utils/ordering.js";
 
 export function sort(comparator, order = asc) {
 	assertFunctor(comparator, `comparator`);
 	assert(order === asc || order === desc, `"order" must be the asc or the desc function.`);
 
-	return function* (observer) {
+	return asObserver(function* (observer) {
+		const cohort = Cohort.from(observer);
 		const items = [];
-
-		let success = true;
 
 		try {
 			while (true) {
 				items.push(yield);
 			}
-		} catch (e) {
-			success = false;
-			observer.throw(e);
+		} catch (error) {
+			abort(cohort, error);
 		} finally {
-			if (success) {
-				items.sort(comparator);
+			close(cohort, () => {
+				if (cohort.isFine) {
+					items.sort(comparator);
 
-				if (order === asc) {
-					for (const item of items) {
-						observer.next(item);
-					}
-				} else {
-					for (let index = items.length - 1; index >= 0; index--) {
-						observer.next(items[index]);
+					if (order === asc) {
+						for (const item of items) {
+							observer.next(item);
+						}
+					} else {
+						for (let index = items.length - 1; index >= 0; index--) {
+							observer.next(items[index]);
+						}
 					}
 				}
-			}
-
-			closeObserver(observer);
+			});
 		}
-	};
+	});
 }
 
 export function sortNumbers(order = asc) {